created separate liveComponent for contract_line
